[refactor] pulls rendering logic out into separate classes
