+++ simulation/src/constants.js
@@ -1,0 +1,4 @@
+export const RENDER_3D = false;
+export const RENDER_SCALE = 1.0;
+export const TILE_SIZE = [384, 384];
+export const TERRAIN_SIZE = [TILE_SIZE[0] * RENDER_SCALE, TILE_SIZE[1] * RENDER_SCALE];
