+++ simulation/src/sim-1.js
@@ -1,8 +1,9 @@
 import "./style/main.css";
 import { vec2, vec3, mat3, mat4 } from "gl-matrix";
 import { DoubleFramebuffer, SingleFramebuffer, DEFAULT_INTERPOLATION } from "./buffer";
-import { DomainMesh } from "./mesh";
 import parameters from './parameters';
+
+import { RENDER_3D, TILE_SIZE, TERRAIN_SIZE } from "./constants";
 
 // assigning regl as a global
 // so that we have access to it in all modules.
@@ -16,10 +17,9 @@ window.regl = require('regl')({
     ]
 });
 
-const RENDER_3D = false;
-const RENDER_SCALE = 1.0;
-const TILE_SIZE = [384, 384];
-const TERRAIN_SIZE = [TILE_SIZE[0] * RENDER_SCALE, TILE_SIZE[1] * RENDER_SCALE];
+const {View3D} = require('./View3d.js');
+const {View2D} = require('./View2d.js');
+const {CrossSection} = require('./CrossSection.js');
 
 const load_image = (url) => {
     return new Promise((accept, reject) => {
@@ -235,23 +235,6 @@ const calculate_stream_averaging = regl({
 
 // end of curvature stuff
 
-// surface normal calculation
-const calculate_N_normals = regl({
-    framebuffer: regl.prop('target'),
-    vert: require('./shaders/pass-through.vert'),
-    frag: require('./shaders/calculate-N-normals.frag'),
-    attributes: {
-        a_position: [[-1, -1], [1, -1], [-1, 1], [1, 1]],
-        a_uv: regl.prop('a_uv')
-    },
-    uniforms: {
-        u_H: regl.prop('u_H'),
-        u_resolution: TILE_SIZE,
-    },
-    primitive: "triangle strip",
-    count: 4
-});
-
 // sediment stuff
 
 const calculate_erosion_accretion = regl({
@@ -373,258 +356,6 @@ const calculate_Q_boundary_conditions = regl({
     primitive: "triangle strip",
     count: 4
 });
-
-
-// GPU calls: rendering
-const render_tile_as_color = regl({
-    vert: require('./shaders/place-tile.vert'),
-    frag: require('./shaders/render-color.frag'),
-    attributes: {
-        a_position: regl.prop('a_position'),
-        a_uv: regl.prop('a_uv')
-    },
-    uniforms: {
-        u_transform: regl.prop('u_transform'),
-        u_color: regl.prop('u_color'),
-        u_resolution: TILE_SIZE
-    },
-    primitive: "triangle strip",
-    count: 4
-});
-
-const render_terrain_height = regl({
-    vert: require('./shaders/place-tile.vert'),
-    frag: require('./shaders/render-terrain-height.frag'),
-    attributes: {
-        a_position: regl.prop('a_position'),
-        a_uv: regl.prop('a_uv')
-    },
-    uniforms: {
-        u_transform: regl.prop('u_transform'),
-        u_H: regl.prop('u_H'),
-        u_scalefactor: regl.prop('u_scalefactor'),
-        u_saturation_point: regl.prop('u_saturation_point'),
-        u_resolution: TILE_SIZE
-    },
-    primitive: "triangle strip",
-    count: 4  
-})
-
-const render_flux = regl({
-    vert: require('./shaders/place-tile.vert'),
-    frag: require('./shaders/render-flux.frag'),
-    attributes: {
-        a_position: regl.prop('a_position'),
-        a_uv: regl.prop('a_uv')
-    },
-    uniforms: {
-        u_transform: regl.prop('u_transform'),
-        u_Q: regl.prop('u_Q'),
-        u_H: regl.prop('u_H'),
-        u_scalefactor: regl.prop('u_scalefactor'),
-        u_resolution: TILE_SIZE
-    },
-    primitive: "triangle strip",
-    count: 4  
-})
-
-const render_flux_magnitude = regl({
-    vert: require('./shaders/place-tile.vert'),
-    frag: require('./shaders/render-flux-magnitude.frag'),
-    attributes: {
-        a_position: regl.prop('a_position'),
-        a_uv: regl.prop('a_uv')
-    },
-    uniforms: {
-        u_transform: regl.prop('u_transform'),
-        u_Q: regl.prop('u_Q'),
-        u_H: regl.prop('u_H'),
-        u_flux_magnitude_scale: regl.prop('u_flux_magnitude_scale'),
-        u_scalefactor: regl.prop('u_scalefactor'),
-        u_resolution: TILE_SIZE
-    },
-    primitive: "triangle strip",
-    count: 4  
-})
-
-const render_slope = regl({
-    vert: require('./shaders/place-tile.vert'),
-    frag: require('./shaders/render-slope.frag'),
-    attributes: {
-        a_position: regl.prop('a_position'),
-        a_uv: regl.prop('a_uv')
-    },
-    uniforms: {
-        u_transform: regl.prop('u_transform'),
-        u_S: regl.prop('u_S'),
-        u_H: regl.prop('u_H'),
-        u_scalefactor: regl.prop('u_scalefactor'),
-        u_resolution: TILE_SIZE
-    },
-    primitive: "triangle strip",
-    count: 4  
-})
-
-const render_curvature = regl({
-    vert: require('./shaders/place-tile.vert'),
-    frag: require('./shaders/render-curvature.frag'),
-    attributes: {
-        a_position: regl.prop('a_position'),
-        a_uv: regl.prop('a_uv')
-    },
-    uniforms: {
-        u_transform: regl.prop('u_transform'),
-        u_H: regl.prop('u_H'),
-        u_Q: regl.prop('u_Q'),
-        u_K: regl.prop('u_K'),
-        u_E: regl.prop('u_E'),
-        u_scalefactor: regl.prop('u_scalefactor'),
-        u_resolution: TILE_SIZE
-    },
-    primitive: "triangle strip",
-    count: 4  
-})
-
-const render_erosion_accretion = regl({
-    vert: require('./shaders/place-tile.vert'),
-    frag: require('./shaders/render-erosion-accretion-values-2.frag'),
-    attributes: { 
-        a_position: regl.prop('a_position'),
-        a_uv: regl.prop('a_uv')
-    },
-    uniforms: {
-        u_transform: regl.prop('u_transform'),
-        u_H: regl.prop('u_H'),
-        u_Q: regl.prop('u_Q'),
-        u_S: regl.prop('u_S'),
-        u_K: regl.prop('u_K'),
-
-        u_k_erosion: regl.prop('u_k_erosion'),
-        u_k_accretion: regl.prop('u_k_accretion'),
-        u_Q_accretion_upper_bound: regl.prop('u_Q_accretion_upper_bound'),
-        u_Q_erosion_lower_bound: regl.prop('u_Q_erosion_lower_bound'),
-
-        u_scalefactor: regl.prop('u_scalefactor'),
-        u_resolution: TILE_SIZE
-    },
-    primitive: "triangle strip",
-    count: 4  
-})
-
-const render_section_line = regl({
-    vert: require('./shaders/place-tile.vert'),
-    frag: require('./shaders/render-section-line.frag'),
-    attributes: {
-        a_position: regl.prop('a_position'),
-        a_uv: regl.prop('a_uv')
-    },
-    uniforms: {
-        u_transform: regl.prop('u_transform'),
-        u_p1: regl.prop('u_p1'),
-        u_p2: regl.prop('u_p2'),
-        u_color: regl.prop('u_color'),
-        u_resolution: TILE_SIZE
-    },
-    primitive: "triangle strip",
-    count: 4 
-});
-
-const render_crosssection = regl({
-    vert: require('./shaders/place-tile.vert'),
-    frag: require('./shaders/render-point-crosssection.frag'),
-    attributes: {
-        a_position: regl.prop('a_position'),
-        a_uv: regl.prop('a_uv')
-    },
-    uniforms: {
-        u_transform: regl.prop('u_transform'),
-        u_p1: regl.prop('u_p1'),
-        u_p2: regl.prop('u_p2'),
-        u_H: regl.prop('u_H'),
-        u_color: regl.prop('u_color'),
-        u_resolution: TILE_SIZE
-    },
-    primitive: "triangle strip",
-    count: 4 
-});
-
-// 3D RENDER CALLS
-let DOMAIN_MESH = new DomainMesh(regl, [512, 512]);
-
-
-const render_domain = regl({
-    framebuffer: null,
-    vert: require('./shaders/place-mesh.vert'),
-    frag: require('./shaders/render-domain.frag'),
-    attributes: {
-        a_position: DOMAIN_MESH.vertices,
-        a_uv: DOMAIN_MESH.uvs,
-        a_id: DOMAIN_MESH.ids
-    },
-    elements: DOMAIN_MESH.indices,
-    uniforms: {
-        u_transform: regl.prop('u_transform'),
-        u_basepoint: regl.prop('u_basepoint'),
-        u_resolution: DOMAIN_MESH.cells,
-
-        u_H: regl.prop('u_H'),
-        u_N: regl.prop('u_N')
-    },
-    primitive: 'triangles',
-    offset: 0,
-    count: DOMAIN_MESH.indices.length * 3.0
-});
-
-const render_river = regl({
-    framebuffer: null,
-    vert: require('./shaders/place-river.vert'),
-    frag: require('./shaders/render-river.frag'),
-    attributes: {
-        a_position: DOMAIN_MESH.vertices,
-        a_uv: DOMAIN_MESH.uvs,
-        a_id: DOMAIN_MESH.ids
-    },
-    elements: DOMAIN_MESH.indices,
-    uniforms: {
-        u_transform: regl.prop('u_transform'),
-        u_basepoint: regl.prop('u_basepoint'),
-        u_resolution: DOMAIN_MESH.cells,
-        u_tex_resolution: TILE_SIZE,
-
-        u_H: regl.prop('u_H'),
-        u_N: regl.prop('u_N'),
-        u_view_pos: regl.prop('u_view_pos')
-    },
-    primitive: 'triangles',
-    offset: 0,
-    depth: { func: 'lequal' },
-    blend: {
-        enable: true,
-        func: {src: 'src alpha', dst: 'one minus src alpha'}
-    },
-    count: DOMAIN_MESH.indices.length * 3.0
-});
-
-// const render_point = regl({
-//     framebuffer: null,
-//     vert: require('./shaders/place-point.vert'),
-//     frag: require('./shaders/render-point.frag'),
-//     attributes: {
-//         a_position: [[-1, -1], [1, -1], [-1, 1], [1, 1]],
-//     },
-//     elements: DOMAIN_MESH.indices,
-//     uniforms: {
-//         u_transform: regl.prop('u_transform'),
-//         u_basepoint: regl.prop('u_basepoint'),
-//         u_H: regl.prop('u_H'),
-//     },
-//     primitive: 'triangle strip',
-//     offset: 0,
-//     count: 4
-// })
-
-// CPU Datastructures
-
 
 
 class Tile {
@@ -898,446 +629,13 @@ class Tile {
     }
 
     render (transform, resources) {
-        // if (this.loaded) {
-        //     // 3D RENDERING STEPS
-        //     if (RENDER_3D) {
-
-        //         // console.log('render');
-        //         regl.clear({depth: 1.0});
-
-        //         let camera = resources.camera;
-        //         // console.log(camera);
-
-        //         const target = camera.target;
-        //         const camera_position = camera.position;
-
-        //         const front = vec3.subtract([], target, camera_position);
-        //         const right = vec3.cross([], front, [0.0, -1.0, 0.0]);
-        //         const up = vec3.cross([], front, right);
-
-        //         const V = mat4.lookAt([], camera_position, target, up);
-        //         const P = mat4.perspective([], Math.PI / 4.0, window.innerWidth / window.innerHeight, 0.001, 100.0);
-
-        //         const PV = mat4.multiply([], P, V);
-
-        //         calculate_N_normals({
-        //             target: this.N.buffer,
-        //             a_uv: this.uvs,
-        //             u_H: this.H.front,
-        //         });
-
-        //         render_domain({
-        //             u_basepoint: [this.x, 0.0, this.y],
-        //             u_transform: PV,
-
-        //             u_H: this.H.front,
-        //             u_N: this.N.buffer
-        //         });
-
-        //         render_river({
-        //             u_basepoint: [this.x, 0.0, this.y],
-        //             u_transform: PV,
-
-        //             u_H: this.H.front,
-        //             u_N: this.N.buffer,
-        //             u_view_pos: camera_position
-        //         });
-
-        //         // render_point({
-        //         //     u_basepoint: [this.x, 0.0, this.y]
-        //         // })
-
-        //     } else {
-        //         // 2D RENDERING STEPS:
-
-        //         regl.clear({depth: 1.0});
-        //         render_terrain_height({
-        //             u_H: this.H.front,
-        //             u_scalefactor: 0.5,
-        //             u_saturation_point: parameters.saturation_point,
-
-        //             a_position: this.positions,
-        //             a_uv: this.uvs,
-        //             u_transform: transform,
-        //         });
-
-        //         if (parameters.render_flux)
-        //         {
-        //             regl.clear({depth: 1.0});
-        //             render_flux({
-        //                 u_Q: this.Q.front,
-        //                 u_H: this.H.front,
-        //                 u_scalefactor: 0.5,
-        
-        //                 a_position: this.positions,
-        //                 a_uv: this.uvs,
-        //                 u_transform: transform,
-        //             })
-        //         }
-
-        //         if (parameters.render_flux_magnitude)
-        //         {
-        //             regl.clear({depth: 1.0});
-        //             render_flux_magnitude({
-        //                 u_Q: this.Q.front,
-        //                 u_H: this.H.front,
-        //                 u_scalefactor: 0.5,
-        //                 u_flux_magnitude_scale: parameters.flux_magnitude_scale,
-        
-        //                 a_position: this.positions,
-        //                 a_uv: this.uvs,
-        //                 u_transform: transform,
-        //             })
-        //         }
-
-        //         if (parameters.render_slope)
-        //         {
-        //             regl.clear({depth: 1.0});
-        //             render_slope({
-        //                 u_S: this.S.buffer,
-        //                 u_K: this.K.front,
-        //                 u_scalefactor: 4.0,
-        
-        //                 a_position: this.positions,
-        //                 a_uv: this.uvs,
-        //                 u_transform: transform,
-        //             })
-        //         }
-                
-        //         if (parameters.render_curvature)
-        //         {
-        //             regl.clear({depth: 1.0});
-        //             render_curvature({
-        //                 u_H: this.H.front,
-        //                 u_K: this.K.front,
-        //                 u_E: this.E.front,
-        //                 u_scalefactor: 4.0,
-        
-        //                 a_position: this.positions,
-        //                 a_uv: this.uvs,
-        //                 u_transform: transform,
-        //             })
-        //         }
-
-        //         if (parameters.render_erosion_accretion)
-        //         {
-        //             regl.clear({depth: 1.0});
-        //             render_erosion_accretion({
-        //                 u_H: this.H.front,
-        //                 u_K: this.K.front,
-        //                 u_Q: this.Q.front,
-        //                 u_S: this.S.buffer,
-
-        //                 u_k_erosion: parameters.erosion_speed,
-        //                 u_k_accretion: parameters.accretion_speed,
-        //                 u_Q_accretion_upper_bound: parameters.accretion_upper_bound,
-        //                 u_Q_erosion_lower_bound: parameters.erosion_lower_bound,
-                        
-        //                 u_scalefactor: 4.0,
-        
-        //                 a_position: this.positions,
-        //                 a_uv: this.uvs,
-        //                 u_transform: transform,
-        //             })
-        //         } 
-
-        //         regl.clear({depth: 1.0});
-        //         render_section_line({
-        //             u_p1: parameters.p1,
-        //             u_p2: parameters.p2,
-        //             u_color: [0.65, 0.2, 0.0],
-        //             a_position: this.positions,
-        //             a_uv: this.uvs,
-        //             u_transform: transform,
-        //         })
-
-        //     }
-
-        // } else {
-        //     console.log('still loading!');
-        //     // If we're still waiting for textures...
-        //     render_tile_as_color({
-        //         a_position: this.positions,
-        //         a_uv: this.uvs,
-        //         u_transform: transform,
-        //         u_color: this.loading_color
-        //     });
-
-        // }        
+              
     }
 
     destroy () {
     }
 }
 
-
-class View2D {
-    constructor(testcase=False) {
-        this.is_testcase = testcase
-
-        this.positions = [
-            [this.x, this.y], [this.x + 1, this.y],
-            [this.x, this.y + 1], [this.x + 1, this.y + 1]
-        ]
-
-        this.uvs = [
-            [0, 0], [1, 0],
-            [0, 1], [1, 1]
-        ]
-
-        this.parent = null
-        this.loaded = false
-        this.loading_color = [0/255, 74/255, 74/255];
-    }
-
-    get_resources() {
-    }
-
-    update_positions () {
-        this.positions = [
-            [this.x, this.y], [this.x + 1, this.y],
-            [this.x, this.y + 1], [this.x + 1, this.y + 1]
-        ]
-    }
-
-    set_parent(parent) {
-        this.parent = parent;
-        this.x = parent.x;
-        this.y = parent.y;
-        this.z = parent.z;
-        this.update_positions();
-    }
-
-    simulate () {}
-
-    render(transform, resources) {
-        if (this.parent.loaded) {
-            // 3D RENDERING STEPS
-            if (RENDER_3D) {
-
-                // console.log('render');
-                regl.clear({depth: 1.0});
-
-                let camera = resources.camera;
-                // console.log(camera);
-
-                const target = camera.target;
-                const camera_position = camera.position;
-
-                const front = vec3.subtract([], target, camera_position);
-                const right = vec3.cross([], front, [0.0, -1.0, 0.0]);
-                const up = vec3.cross([], front, right);
-
-                const V = mat4.lookAt([], camera_position, target, up);
-                const P = mat4.perspective([], Math.PI / 4.0, window.innerWidth / window.innerHeight, 0.001, 100.0);
-
-                const PV = mat4.multiply([], P, V);
-
-                calculate_N_normals({
-                    target: this.parent.N.buffer,
-                    a_uv: this.parent.uvs,
-                    u_H: this.parent.H.front,
-                });
-
-                render_domain({
-                    u_basepoint: [this.x, 0.0, this.y],
-                    u_transform: PV,
-
-                    u_H: this.parent.H.front,
-                    u_N: this.parent.N.buffer
-                });
-
-                render_river({
-                    u_basepoint: [this.x, 0.0, this.y],
-                    u_transform: PV,
-
-                    u_H: this.parent.H.front,
-                    u_N: this.parent.N.buffer,
-                    u_view_pos: camera_position
-                });
-
-                // render_point({
-                //     u_basepoint: [this.x, 0.0, this.y]
-                // })
-
-            } else {
-                // 2D RENDERING STEPS:
-
-                regl.clear({depth: 1.0});
-                render_terrain_height({
-                    u_H: this.parent.H.front,
-                    u_scalefactor: 0.5,
-                    u_saturation_point: parameters.saturation_point,
-
-                    a_position: this.positions,
-                    a_uv: this.uvs,
-                    u_transform: transform,
-                });
-
-                if (parameters.render_flux)
-                {
-                    regl.clear({depth: 1.0});
-                    render_flux({
-                        u_Q: this.parent.Q.front,
-                        u_H: this.parent.H.front,
-                        u_scalefactor: 0.5,
-        
-                        a_position: this.positions,
-                        a_uv: this.uvs,
-                        u_transform: transform,
-                    })
-                }
-
-                if (parameters.render_flux_magnitude)
-                {
-                    regl.clear({depth: 1.0});
-                    render_flux_magnitude({
-                        u_Q: this.parent.Q.front,
-                        u_H: this.parent.H.front,
-                        u_scalefactor: 0.5,
-                        u_flux_magnitude_scale: parameters.flux_magnitude_scale,
-        
-                        a_position: this.positions,
-                        a_uv: this.uvs,
-                        u_transform: transform,
-                    })
-                }
-
-                if (parameters.render_slope)
-                {
-                    regl.clear({depth: 1.0});
-                    render_slope({
-                        u_S: this.parent.S.buffer,
-                        u_K: this.parent.K.front,
-                        u_scalefactor: 4.0,
-        
-                        a_position: this.positions,
-                        a_uv: this.uvs,
-                        u_transform: transform,
-                    })
-                }
-                
-                if (parameters.render_curvature)
-                {
-                    regl.clear({depth: 1.0});
-                    render_curvature({
-                        u_H: this.parent.H.front,
-                        u_K: this.parent.K.front,
-                        u_E: this.parent.E.front,
-                        u_scalefactor: 4.0,
-        
-                        a_position: this.positions,
-                        a_uv: this.uvs,
-                        u_transform: transform,
-                    })
-                }
-
-                if (parameters.render_erosion_accretion)
-                {
-                    regl.clear({depth: 1.0});
-                    render_erosion_accretion({
-                        u_H: this.parent.H.front,
-                        u_K: this.parent.K.front,
-                        u_Q: this.parent.Q.front,
-                        u_S: this.parent.S.buffer,
-
-                        u_k_erosion: parameters.erosion_speed,
-                        u_k_accretion: parameters.accretion_speed,
-                        u_Q_accretion_upper_bound: parameters.accretion_upper_bound,
-                        u_Q_erosion_lower_bound: parameters.erosion_lower_bound,
-                        
-                        u_scalefactor: 4.0,
-        
-                        a_position: this.positions,
-                        a_uv: this.uvs,
-                        u_transform: transform,
-                    })
-                } 
-
-                regl.clear({depth: 1.0});
-                render_section_line({
-                    u_p1: parameters.p1,
-                    u_p2: parameters.p2,
-                    u_color: [0.65, 0.2, 0.0],
-                    a_position: this.positions,
-                    a_uv: this.uvs,
-                    u_transform: transform,
-                })
-
-            }
-
-        } else {
-            console.log('still loading!');
-            // If we're still waiting for textures...
-            render_tile_as_color({
-                a_position: this.positions,
-                a_uv: this.uvs,
-                u_transform: transform,
-                u_color: this.loading_color
-            });
-
-        }
-    }
-}
-
-
-// CrossSection
-
-class CrossSection {
-    constructor(testcase=False) {
-        this.is_testcase = testcase
-
-        this.positions = [
-            [this.x, this.y], [this.x + 1, this.y],
-            [this.x, this.y + 1], [this.x + 1, this.y + 1]
-        ]
-
-        this.uvs = [
-            [0, 0], [1, 0],
-            [0, 1], [1, 1]
-        ]
-
-        this.parent = null
-        this.loaded = false
-    }
-
-    get_resources() {}
-
-    update_positions () {
-        this.positions = [
-            [this.x, this.y], [this.x + 1, this.y],
-            [this.x, this.y + 1], [this.x + 1, this.y + 1]
-        ]
-    }
-
-    set_parent(parent) {
-        this.parent = parent;
-        this.x = parent.x + 1.0;
-        this.y = parent.y + 0.0;
-        this.z = parent.z;
-        this.update_positions();
-    }
-
-    simulate () {}
-
-    render(transform, resources) {
-        if (this.parent == null || !this.parent.loaded) { return; }
-        if (RENDER_3D) { return; }
-
-        regl.clear({depth: 1.0});
-        render_crosssection({
-            u_p1: parameters.p1,
-            u_p2: parameters.p2,
-            u_H: this.parent.H.front,
-
-            a_position: this.positions,
-            a_uv: this.uvs,
-            u_transform: transform,
-            u_color: this.loading_color
-        });
-
-    }
-}
 
 // TileProvider
 class Camera {
@@ -1371,12 +669,14 @@ class TileProvider {
             // Real DEM Stuff
             // new Tile(0, 0, 14),
             new View2D(true),
+            new View3D(true),
             new CrossSection(true)
         ];
 
         // hook up the cross section renderer
         this.tiles[1].set_parent( this.tiles[0] );
         this.tiles[2].set_parent( this.tiles[0] );
+        this.tiles[3].set_parent( this.tiles[0] );
 
         this.tile_map = {};
         this.resources = { 
@@ -1416,9 +716,9 @@ class TileProvider {
         let s = performance.now();
 
         // regl.clear({color: [0, 0, 0, 1]});
-        this.tiles.forEach((tile, i) => { 
-            tile.simulate(this.transform, this.resources);
-            tile.render(this.transform, this.resources); 
+        this.tiles.forEach((tile) => { 
+            tile.simulate(this.transform, this.resources, parameters);
+            tile.render(this.transform, this.resources, parameters); 
         });
         this.resources.t += (parameters.running) ? 1 : 0;
